e2e: remove multi-apps test + fix frontend source metrics names (#2827)
## Description

This test is redundant with the Source e2e, except for:

1. Verify node odiglet label added
2. Verify collector metrics in the UI

These 2 can also probably be de-duplicated but this is a good first step

Update: This also fixes a bug in the frontend's source metrics service
which stores metrics based on their _workload_ name, but retrieves them
based on their _service_ name, which means when a source has a custom
service name, its throughput metrics won't get retrieved by the graphql
query. This was caught when moving the collector metrics UI case because
the Source e2e sets a custom service name.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [x] Updated e2e Tests
- [ ] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly