Add dependencies-syncer-bot PR check to release workflow (#2837)
## Description

Similar to https://github.com/odigos-io/odigos/pull/2680, this adds a
pre-release check for open enterprise dependency sync PRs.

Currently, the release flow will create a release PR in the enterprise
repo that is effectively a dependency sync. However, if there are
currently failing tests with the current dependency sync, it would be
good to catch these before we start a release flow instead of having to
debug it mid release.

Making this a separate job so it can depend on the verify-offsets check,
which itself will trigger a dependency bump if there are offsets changes
to be merged

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [ ] Updated e2e Tests
- [ ] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly