fix(helm): do not default mountMethod to allow profile override (#2840)
If user did not set `mountMethod` in helm values, do not default it in
helm.

This is to accurately differentiate between user did not set any value,
and user set a value to be the `k8s-virtual-device` (default).

odigos profile will only override the value if it's empty (assuming that
if a user specified a value it will be used and not the profile), so
this change allows us use profiles for mount method in helm correctly