feat: add support for LD_PRELOAD odigos-loader (#2759)
Using LD_PRELOAD env var as a way of adding the instrumentation relevant
environment variables.
This feature will be controlled by a configuration flag. A new
configuration called `envVarInjectionMethod` is added which have the
values
1. `pod-manifest` indicating we should only try to add env vars through
the pod manifest (what we do today). This is the default value.
2. `loader-fallback-to-pod-manifest` - first try using the loader - if
that is not possible, fallback to the pod-manifest approach. Reasons for
that are: a) the user have an already defined LD_PRELOAD value - and
currently we want to avoid manipulating that value. b) we detect secure
execution mode in the container.
3. `loader` - only trying the loader method. If we can't add it for one
of the reasons mentioned we won't instrument the pod.