feat: add k8s config map provider to collector (#2842)
Replacing the `odigosfilleprovider` with `odigosk8scmprovider`.
There are 2 reasons for this change:
1. The old approach required `fsnotify` which used the underlying
machine's inotify quota. Since we have a daemonset collector - this
quota was often reached and cause the collector to crash - for nodes
that were low on this resource or set a low amount.
2. The fileprovider approach has a builtin latency that depends on the
kubelet sync period. This means that the time it takes for the collector
to update its config with that approach can be up to a minute.

This PR is adding the new `k8scmprovider` which is using a cache
`Informer` to watch the provided config map for updates. Thus, removing
the use of `fsnotify` and makes the updates pretty much immediate.
Adding Roles for the node and cluster collectors to watch and list the
corresponding config maps - updated in helm and CLI installations.
The autoscaler is updated to use the `--conf` flag in a way that
utilizes the new provider:
```
  containers:
  - command:
    - /odigosotelcol
    - --config=k8scm:odigos-system/odigos-data-collection/conf
    
    ....

  containers:
  - command:
    - /odigosotelcol
    - --config=k8scm:odigos-system/odigos-gateway/collector-conf
    
```
The config map mounts and volumes are removed.

This was tested manually, and in addition we have a test phase that
asserts the collector log to indicate a config update.
Adding observability for the collector config updates is something we
should add in the future and there are few options of doing that.
Currently the collector doesn't seem to support adding metrics for a
config provider similar to what is supported in other components.