fix: handle empty path in url templating (#2864)
## Description

When the url path is empty string like in "http://example.com", the span
names ends up like "GET " (extra space at the end), which:
- creates 2 different values in UIs (both "GET" and "GET " with space),
which is confusing to users.
- non intuitive what it actually means.

This PR fixes the behavior so that empty values in http span names are
always set to "/" to denote root path.

## How Has This Been Tested?


- [x] Added Unit Tests
- [ ] Updated e2e Tests
- [ ] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

No

## User Facing Changes

Fix bug in processor which user will notice