add uninstall test, upgrade simple-demo test apps (#2861)
* Adding uninstall step to the `sources` e2e test. This test includes
verifying that: all Odigos CRDs are removed, all the instrumented pods
are deployed without Odigos modifications and all of our config maps and
secrets are removed.
* This test uncovered two minors bugs: 
1) the `effective-config` config map was not removed properly (in Helm
mainly) - since it is not created by helm but by
our controller we should make sure it is cleaned. This is done by adding
an owner reference to the `odigos-config`
config map which is created by Helm/CLI. 
2) A race condition in uninstall that may lead to un-removed sources -
this was handled by checking for un-deleted
sources during the polling phase.
* Updating the `simple-demo` applications used in the tests to the
latest version. This mainly includes a change to the `currency`
deployment that now uses a multi container pod. The assertions are
updated to verify that the runtime detection is done correctly.