feat: use cert-controller for certificates rotation and creation (#2876)
Changing the way we manage the certificates for our webhooks - making
use of
[cert-controller](https://github.com/open-policy-agent/cert-controller)
to manage the certificates in the controller instead of doing it in
CLI/Helm.

### Issues with our current approach
Today our certificates are managed by CLI/helm, this has the following
issues:
* The certificates are refreshed every time an `install`, `upgrade`,
`profile`, `config` or `helm upgrade` commands are executed. This is
both not necessary and causes the webhooks to have a period of 30-100s
of down time - this is a result of the kubelet sync period
(https://ahmet.im/blog/kubernetes-secret-volumes-delay/). The
implications are that after one of the mentioned commands are executed
we'll fail to create sources and instrument newly created pods.
* The certificate rotation mechanism is not deterministic and depends on
when the commands are running, in theory a cluster which has odigos
installed for the duration of the certificate validity will become
un-usable once the certificate is expired.
* Having this managed by the CLI and helm can cause mismatches when
Odigos is installed with Helm and CLI commands are ran - we saw this
issue because the CA name in the CLI was `odigos-instrumentor` and in
helm `sserving-cert` this caused bad certificate errors.
* We have the `crypto` package in the CLI that handles the certificates
creation, this is not ideal since we don't have a mechanism to detect
security patches required for it.

### Using `cert-controller`
* The `cert-controller` packages allows to register a controller that
watches the certificate secret and the webhooks configurations and
aligns the certificate and the `caBundle` on the webhooks. It has a
configurable period of 12 hours to check if the certificates need
rotation (about to expire) - it can be configured with
`LookaheadInterval`.
* Using the new package also for the initial certificate creation allows
us to remove this from the CLI and Helm and remove any possible
mismatch.
* The package provides a `IsReady` channel which we use to delay the
registration of the webhooks. The modified init flow is:
1. Create the manager.
2. Register `cert-controller` with the manager.
3. Start the manager (in a goroutine)
4. Wait for the IsReady signal (in a goroutine) - after that has arrived
-> register the webhooks -> mark the readiness probe to return true.
This flow will guarantees that the pod will be marked ready only after
the webhooks are usable.
It is used in the instrumentor and the autoscaler. Added readiness probe
for the autoscaler.

### Cons with the new approach
* New permissions are required for the instrumentor/autoscaler pods -
watching and listing webhook configurations and watching and updating
the certificate secret. These permissions are as minimal as possible -
limited to our namespace and specific resources to update - to minimize
the permission additions.
* Once the certificate is rotated, the kubelet sync period will cause
the same affect. With this change this will happen `LookaheadInterval`
before the cert is expired. Assuming 1 year expiration and 100 sec
kubelet sync period that is 99.99968% availability - for more details:
https://github.com/open-policy-agent/cert-controller/issues/50
* The `certRotator` should probably only be launched in the leader
instance, but this is not possible to do in a reliable way due to
https://github.com/open-policy-agent/cert-controller/issues/53. So
currently we have the cert controller running on both leader and non
leader instances. This can lead to conflict error if the 2 instances are
trying to update the same resource at the same time - but the reconcile
retry logic will handle this.