feat: run cluster collector before destinations are added (#2738)
## Description

At the moment, cluster collector is only started once a destination is
added. This causes unnecessary delay in demos and ramp-ups.

With this change, scheduler will always create the cluster collector
group right after effective config is calculated. then autoscaler will
pick this up and deploy the cluster collector.

By the time the user opens the UI and adds its sources, we only need to
create node collector which will be faster.

In the future we might want to add a mode for odigos that only sets up
instrumentation but allows users to send to their own collector and not
deploy our cluster collector. in this case this logic should be
modified.

## How Has This Been Tested?

- [x] Manual Testing

## Kubernetes Checklist

Removed RBAC permissions for scheduler as it's no longer used after this
change 🚀

- [x] Changes how Odigos interacts with Kubernetes
- [x] Modifies Odigos manifests (addressed in both CLI and Helm)
- [x] Changes RBAC permissions

## User Facing Changes

After installing odigos, an additional pod for cluster collector will
run even before adding any destination. this should not be dramatic as
it's just one more deployment to the 4 that we already deploy and the
ds. plus it's only for a short while as the collector will start anyway
once the user setup odigos in ui or gitops

---------

Co-authored-by: Ben Elferink <ben.elferink@icloud.com>