feat: enable metrics server in Odiglet when not using hostNetwork. Add MeterProvider for custom metrics. Add ebpf manager metrics. (#2931)
## Description

Following #2928, we can safely expose metrics server in the odiglet. For
k8s versions >= 1.26 we are running without `hostNetwork` - and we can
bind to a const and known port without handling conflicts. Using the
default metrics port used by controller-runtime.
controller-runtime has a builtin metrics server using prometheus - it
includes metrics for all the controllers and the go runtime.
This PR **does not aim** to scrape the metrics, and this is left for
follow ups. The goal in this PR is to setup the metrics server and add
`MeterProvider` for custom metrics.
We are setting up an OTel `MetreProvider` that is configured to export
in prometheus format and integrates with the prometheus metric registry
exposed by controller-runtime.
The `MeterProvider` can be passed to different components in the
Odiglet, each can declare its own specific metrics.
Adding the `MeterProvider` to the eBPF instrumentation manager as an
initial step - including 2 metrics: first for the number of currently
instrumented processes, and the second for the number of processes we
failed to instrument.
In the future the `MeterProvider` can be passed and used by the `OpAmp`
server, the `runtime-detector` and the different instrumentations
libraries - each can add specific metrics which will end up in the
metrics server.