fix(frontend): retry create object with generated name on conflict (#2974)
## Description

see
https://github.com/kubernetes/enhancements/blob/master/keps/sig-api-machinery/4420-retry-generate-name/README.md

> But there is a problem with name conflicts. There is a 50% chance of
hitting a conflict before generating even 5000 names, and a 0.1% chance
at only 500 names!
> 
> Today a HTTP 409 respose returned to the client if a name conflict
happens. It is possible for clients to detect the error and retry, but
we've seen that most clients don't realize to do this and hit this
problem in production.

It appears that pre k8s 1.32, every call to create a new resource in k8s
with `GenerateName` can fail with this error reported from one of our
users:


![image](https://github.com/user-attachments/assets/bd2933a3-04af-4f35-b275-d48feaad7ae2)

This PR fixes this bug by retrying the Create operation if we got back
conflict and the k8s version is < 1.32.

With k8s >= 1.32, the api server will automatically retry the create if
it gets a conflict.


## How Has This Been Tested?

Hard to reproduce this in a deterministic way. I run this locally and
verified everything still works.

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [x] Changes how Odigos interacts with Kubernetes
- [x] Introduces additional calls to the API Server (potential
performance impact)
retry mechanism recommended by k8s team. should happen very rarely
- [x] New Query/feature supported in all the k8s versions supported by
Odigos
         added feature gate

## User Facing Changes

The above error should not happen anymore