feat: add agent injected to describe (#3001)
Add more visibility for describe (soon in UI @BenElferink )

Per Pod:
- "Latest Workload Revision": shows true/false based on if the pod is
running latest revision of the deployment. e.g. if we are in rollout in
progress, it will help identify pods that are pre-rollout and are not
relevant. Only works for deployment at the moment (no daemonset or
statefulset).
- "Odigos Agent Injected": wether this pod is marked by the
"odigos.io/agents-meta-hash" annotation. e.g. that the pods webhook run
for this pod and applied the relevant setup for instrumentation. it
means env / mounts were applied, but agent may still fail to start later
on. this property can list out pods that are not reporting data since
they were not handled by the webhook.

Per Container in Pod:
- "Started"
- "Ready"
Reflect the status on the pod itself, helps to verify and troubleshoot
issues. While the pod phase might show "Running", individual containers
can fail to start or take long time to spin-up. this info helps identify
when pods has issues starting to set expectation for telemetry, or
signal possible crash issues that may or may not be related to
instrumentation agent.


![image](https://github.com/user-attachments/assets/62cbfc88-bca8-43ff-aab6-817fac60193a)