feat: add option to disable automatic rollout of workloads in odigos (#3022)
## Description


Some user perfer to manage the process of rolling out their workloads,
and not have odigos perform automatic rollouts when instrumenting and
uninstrumenting.

This PR adds a new option to config, which allows to disable the
automatic rollout.
If this feature is used, users are expected to rollout their pods
themselves in order to have or remove instrumentation.

The new feature in UI now shows, per pod, if odigos agent is injected to
the pod or not.

## How Has This Been Tested?

- [x] Manual Testing

## Kubernetes Checklist

Modifies the conditions of instrumentation config

## User Facing Changes

Currently not documented in the docs website, until the feature is
complete with more visibility into what's going on and better fine
grained conditions.

---------

Co-authored-by: Tamir David <tamir@odigos.io>