Dynatrace agent detection all lang - CloudNative mode (#3040)
## Description
This PR introduces detection logic for existing Dynatrace agent
installations in the cluster. The goal is to prevent conflicts when
Odigos instruments user applications.
 
Dynatrace supports 2 deployment modes, and this changes covers all
languages except php when working with `CloudNativeFullStack`, and for
python apps in `ClassicFullStack` mode.
This apparoach relies on `LD_PRELOAD` environment variable the agent
inject into the app with specific values it uses.

<!-- Short summary of the changes -->
The detection mechanism relies on inspecting the LD_PRELOAD environment
variable, which Dynatrace injects into application processes with known
path patterns (e.g., /opt/dynatrace/oneagent...). By identifying these
values, we can infer the presence of the Dynatrace agent and skip
instrumentation as needed.
## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->
kv-mall apps on eks cluster

![Screenshot 2025-06-18 at 12 26
39](https://github.com/user-attachments/assets/390e1aa8-e5bc-4e86-a731-76a2e63289c6)
![Screenshot 2025-06-18 at 13 34
08](https://github.com/user-attachments/assets/8d3dc37c-88ea-40e7-9576-cd52944e6074)

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly