installed Datadog agent detection  (#3047)
## Description
Following the same approach as [this
PR](https://github.com/odigos-io/odigos/pull/3040), we want the ability
to check if Datadog is already installed in the cluster in order to
avoid conflicts with Odigos instrumentation.
 
<!-- Short summary of the changes -->
We look for a specific environment variable that the Datadog agent
injects into the pod. If it exists, we propagate it to the `otherAgent`
configuration.

This detection logic been checked and works well for all supported
languages except PHP.
## How Has This Been Tested?
Manual testing

![Screenshot 2025-06-19 at 16 45
34](https://github.com/user-attachments/assets/04319727-2eba-4fd1-a7ee-5121acc8d93d)