chore: remove unused deps from nodejs agent node_modules (#3058)
For the nodejs agent , we currently run `yarn install` and copy the
generated `node_modules` into the odiglet.

This brings in all the dev dependencies, which includes typescript with
21 MB on disk.

After this change, we install only production dependencies, and can
reduce 21 MB from the folder size we need to embed in odiglet image, and
we need to use on the host.