refactor(instrumentor): fetch the object once in sync function instead of each caller (#3064)
## Description

The `syncWorkload` function in the `source instrumentation` controllers
is called 3 times, and every caller needs to run the exact same lines:

```
	obj := workload.ClientObjectFromWorkloadKind(objKind)
	err := k8sClient.Get(ctx, key, obj)
	if err != nil {
		return ctrl.Result{}, client.IgnoreNotFound(err)
	}
```

We already have the ctx and k8s client in the function, so better to do
it there just once.

This one also guarantee that we are always working with a fresh copy
from cache, and not something we `List`ed maybe some time ago and during
the iteration became invalid.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [x] Manual Testing

## Kubernetes Checklist

No K8s changes. We are still making the same calls but now in one
central place. All calls are to the cache and not api server

## User Facing Changes

No user changes, just internal refactor