refactor(instrumentor): move svc name update to source controller (#3067)
Currently we are setting the initial instrumentation config in
`sourceinstrumentation` controllers, but update it in
`instrumentationconfig` controllers.

Since we are already updating the instrumentation config in the
`sourceinstrumentation` controllers (for labels and container
overrides), this simplifies our architecture a bit, grouping similar
things in one place, and make sure we have less api calls to k8s

## Test

tested it manually with various scenarios to make sure `serviceName` in
`instrumentation config` is updated when changed

## K8s

This will make sure that any updates we need to do to instrumentation
config based on source CRD change are happening in one place and as
single operation. Should reduce a bit our api calls (or at least not
increase them).

This PR removes one controller-runtime controller from odigos (merging
it's functionality into existing controller).