refactor(instrumentor): use distro instead of otelsdk to inject signal envs (#3070)
## Description

There is an ongoing effort to migrate from the "OtelSDK" notation to the
distro notation, managing properties in distro yaml files instead of
doing it with "if"s in code in dozens of places.

This PR is another step in the way, migrating the "static signals otel
env vars" to derive and be calculated as part of distro flow.

- a new property is added to distro called "signalsAsStaticOtelEnvVars"
- set this property to `true` for community `java`, `php`, `ruby` which
were set to use it in code.
- When this property is set to `true`, the pods webhook will inject
`OTEL_LOGS_EXPORTER`, `OTEL_METRICS_EXPORTER`, `OTEL_TRACES_EXPORTER`
with values either `otlp`, or `none` based on the enabled signals in the
node collectors group.
- notice that these env vars are static for the time of pod creation.
this was the case previously, but it's encouraged to dynamically
propagate these values into the agent so to have all instrumented
signals collected correctly at all times.

I am still fetching the collectors group. now instead of doing this in
internal function, it is done in the main webhook call. In the future,
we should write the enabled signals per container and avoid using the
collectors group object in the webhook.


## How Has This Been Tested?

- [x] Manual Testing
- [x] e2e testing - added a missing assert for these env vars in e2e for
frontend. php and ruby already have it @BenElferink great job 💪

## Kubernetes Checklist

We are now always `Get`ing the collectors group object from Cache, even
for languages that do not need it. this is temporary and will be
refactored in a followup PR.

## User Facing Changes

No