fix: inject agent only if signals are enabled (#3097)
## Description

Since https://github.com/odigos-io/odigos/pull/2738 , the cluster
collectors group is generated even if there are no destinations.
That means that once sources are added, the datacollection will become
ready and we will inject agent and rollout.

This is problematic because:

- Some distributions consumes the enabled signals as env vars at pod
creation time, and currently we are rollouting too early, creating pods
with all signals disabled which are not being updated later once a
destination is added.
- We are creating noise in the cluster, doing rollouts before it's
relevant.
- There is no indication to the user that the agent is injected but all
signals are disabled, thus they should not expect any telemetry

Fix: add a new condition for that, and test it in agentenabled
controller of instrumentor.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [x] Manual Testing

## Kubernetes Checklist

No

## User Facing Changes

Show up like this in the ui:


![image](https://github.com/user-attachments/assets/57435db7-3c72-43fd-9f71-762fe4b76c8a)