fix(instrumentor): predicate for agent injection on rules check distros (#3098)
## Description

Fix 2 bugs in the triggering of agent injection from instrumentation
rules changes:

1. the predicate only checks for the deprecated "OtelSdk" rule, and not
checking for the new "OtelDistros" rule. this is fixed to check both
2. the reconciler only trigger recalculate if the rule is not `nil`, but
it need to cleanup any possible modification even if the rule is nil if
before the update it was not nil.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [ ] Updated e2e Tests
- [x] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

More reconciles invocation to catch events that we should have handled
and missed before.

## User Facing Changes

If any user is using the otelDistro rule (which is not yet in the UI)
then his changes will take effect where before they did not immediately
did.