fix(instrumentor): apply trace config rules to agent enabled signals (#3101)
## Description

The instrumentation config containers array in the spec contains enabled
signals, which are currently used for injecting env vars for collectors
in java ruby and php.
However, this list only takes into effect the signals enabled in
collector, but any instrumentation rule that disables traces is not
calculated here.

This PR fixes that, so if someone configures an instrumentation rule for
disabling traces, the fields in the container agent config will reflect
this correctly.

It also make it so if a rule disables traces so that no more signals are
left, we will avoid enabling the agent and show a good message
indicating what is going on:


![image](https://github.com/user-attachments/assets/6d147948-3522-4cf0-aee6-a13216e6af50)


![image](https://github.com/user-attachments/assets/3cc675aa-e34d-4cc7-9886-8678aa193fc5)


## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [x] Manual Testing

## Kubernetes Checklist

No

## User Facing Changes

One more option to see as to why agent is not enabled in a specific
source