refactor: dotnet agent make directory name arch agnostic (#3102)
## Description

dotnet used to have directories containing the arch suffix
(linux-glibc-arm64).
this works will with virtual device that knows the arch it is running
on.
however, the webhook cannot know in advance which arch the pod is going
to run on.
thus, the directory names are renamed so they do not contain the arch
suffix (linux-glibc)
which can be used by the webhook.

---------

Co-authored-by: Tamir David <tamir@odigos.io>