refactor(opamp): add ic predicates to reduce events (#3109)
## Description

Opamp server had no predicate on the instrumentation configs it was
processing.
This caused us to reconcile too much, where we can limit the number of
events like we do in other places.

This PR adds trivial filter which should already reduce quite a lot of
events:
- generation change - filter condition updates and runtime details
detections

The filter can be refined further, but I want to start with these and
improve the filtering as I refactor the mechanism later on.

## How Has This Been Tested?

- [x] Manual Testing

## Kubernetes Checklist

Odiglet will reconcile less - less CPU. no effect on API server.

## User Facing Changes

Internal refactor