Argocd docs renewal (#3111)
## Description
This PR updates the documentation for deploying Odigos using Argo CD and
the official Helm chart. It provides a streamlined onboarding guide,
including a complete Application manifest example with the necessary
configuration values.

<!-- Short summary of the changes -->
With the transition to using a pod-injecting webhook, configuring
ignoreDifferences in Argo CD is no longer required. This page now
focuses solely on the manual setup of Odigos through Argo CD and Helm,
offering a simpler and more guided installation experience.

As a final refinement, the outdated kubectl port-forward command has
been replaced with the simpler odigos-ui CLI command for launching the
UI post-installation.

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly