operator: fix broken UIModeNormal removal from CRD enum (#3124)
## Description

Cursor pointed out
https://github.com/odigos-io/odigos/pull/3121#pullrequestreview-2975588158

The `normal` value was removed from UIMode options. This is fine for
every other install mode, but in the operator it will break existing
CRDs. This adds `normal` to the available enum and adds a translation
check for existing `normal` CRDs to move to `default`.

This also adds a test for the Operator CRD to be generated correctly
from enum values like this

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [ ] Updated e2e Tests
- [ ] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly